Simplified functions that return history.
A history is now just a list of Revisions. A Revision now contains
a list of files modified in that revision.  parseGitLog returns a
list of Revisions.  The old data structure for git log entries, and
the function that converts these to History, are no longer needed
and have been removed.  So has the History type synonym.