Use --max-count=1 if possible in darcsLatestRevId
Without --max-count=1, we have to retrieve and parse
the whole repo history just to get the ID of the latest
change touching a file. This is incredibly inefficient.

Unfortunately, --max-count=1 is currently a darcs darcs
feature only. So we provide a fallback:  if
'darcs changes --max-count=1' fails, we run 'darcs changes'
again without --max-count=1.