Data.FileStore.Utils: +grep flag to ignore binaries
The -I flag is useful because if there is a binary inside the repo,
then grep will say something like 'binary foo matches', which breaks
our little grep parser, which is expecting something just like
'file:56:this is a matching line'. But -I will cause all the binary
matches to be suppressed, fixing the issue.